+++ exporter/opentelemetry-exporter-prometheus/tests/__init__.py
@@ -1,4 +1,3 @@
-# Copyright 2019, OpenCensus Authors
 # Copyright The OpenTelemetry Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,5 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-__version__ = "1.10.0"